Refactor code to have all AlamPropsData logic inside Alarm class. Remove AlarmPros struct.
